Implement three-line lists.
