+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         tools:ignore="GoogleAppIndexingWarning">
 
         <activity
-            android:name=".ui.EntryActivity"
+            android:name=".ui.activities.EntryActivity"
             android:noHistory="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,18 +27,18 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".ui.LauncherActivity" android:launchMode="singleTop" />
-        <activity android:name=".ui.WelcomeActivity" />
-        <activity android:name=".ui.ProfileActivity"
+        <activity android:name=".ui.activities.LauncherActivity" android:launchMode="singleTop" />
+        <activity android:name=".ui.activities.WelcomeActivity" />
+        <activity android:name=".ui.activities.ProfileActivity"
             android:label="@string/app_name"
-            android:parentActivityName=".ui.LauncherActivity">
+            android:parentActivityName=".ui.activities.LauncherActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ui.LauncherActivity" />
+                android:value=".ui.activities.LauncherActivity" />
         </activity>
-        <activity android:name=".ui.SettingsActivity" android:parentActivityName=".ui.LauncherActivity">
+        <activity android:name=".ui.activities.SettingsActivity" android:parentActivityName=".ui.activities.LauncherActivity">
             <meta-data android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.example.dsychyov.sychyovmd.ui.LauncherActivity" />
+                android:value="com.example.dsychyov.sychyovmd.ui.activities.LauncherActivity" />
         </activity>
 
         <meta-data android:name="io.fabric.ApiKey"
