+++ app/src/main/java/com/example/dsychyov/sychyovmd/ui/activities/BaseActivity.java
@@ -1,8 +1,10 @@
-package com.example.dsychyov.sychyovmd.ui;
+package com.example.dsychyov.sychyovmd.ui.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.example.dsychyov.sychyovmd.ui.Utils;
 
 public abstract class BaseActivity extends AppCompatActivity {
