+++ app/src/main/java/com/example/dsychyov/sychyovmd/ui/fragments/ProfileItemsFragment.java
@@ -10,7 +10,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.dsychyov.sychyovmd.ui.OffsetItemDecoration;
 import com.example.dsychyov.sychyovmd.R;
 import com.example.dsychyov.sychyovmd.models.ProfileItem;
 import com.example.dsychyov.sychyovmd.ui.adapters.ProfileItemsAdapter;
@@ -34,10 +33,6 @@ public class ProfileItemsFragment extends Fragment {
             Context context = view.getContext();
             RecyclerView recyclerView = (RecyclerView) view;
 
-            recyclerView.setHasFixedSize(true);
-            final int offset = getResources().getDimensionPixelSize(R.dimen.launcher_icon_margin);
-            recyclerView.addItemDecoration(new OffsetItemDecoration(offset));
-
             recyclerView.setLayoutManager(new LinearLayoutManager(context));
             recyclerView.setAdapter(new ProfileItemsAdapter(prepareProfileInfo()));
         }
@@ -48,13 +43,13 @@ public class ProfileItemsFragment extends Fragment {
     private List<ProfileItem> prepareProfileInfo() {
         List<ProfileItem> list = new ArrayList<>();
 
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_dialog_info), getString(R.string.author_name), ""));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_menu_call), "(33) 652 64 09", ""));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_dialog_email), getString(R.string.author_email), "Email"));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_menu_share), "vk.com/denissychyov", "VK"));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_menu_share), "facebook.com/denis.sychou", "Facebook"));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_menu_share), "steamcommunity.com/id/gost402", "Steam"));
-        list.add(new ProfileItem(getResources().getDrawable(android.R.drawable.ic_menu_share), "github.com/gost402", "Github"));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_info), getString(R.string.author_name), getString(R.string.profile_author)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_phone), "(33) 652 64 09", getString(R.string.profile_phone)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_email), getString(R.string.author_email), getString(R.string.profile_email)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_browser), "vk.com/denissychyov", getString(R.string.profile_vk)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_browser), "facebook.com/denis.sychou", getString(R.string.profile_facebook)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_browser), "steamcommunity.com/id/gost402", getString(R.string.profile_steam)));
+        list.add(new ProfileItem(getResources().getDrawable(R.drawable.ic_action_browser), "github.com/gost402", getString(R.string.profile_github)));
 
         return list;
     }
