+++ app/src/main/res/layout/activity_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.dsychyov.sychyovmd.ui.ProfileActivity">
+    tools:context="com.example.dsychyov.sychyovmd.ui.activities.ProfileActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
@@ -45,7 +45,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="com.example.dsychyov.sychyovmd.ui.ProfileActivity"
+        tools:context="com.example.dsychyov.sychyovmd.ui.activities.ProfileActivity"
         tools:showIn="@layout/activity_profile">
 
         <fragment
