+++ app/src/main/res/layout/activity_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.dsychyov.sychyovmd.ui.WelcomeActivity"
+    tools:context="com.example.dsychyov.sychyovmd.ui.activities.WelcomeActivity"
     android:layout_margin="@dimen/welcome_page_margin"
     android:orientation="vertical">
