+++ app/src/main/res/layout/fragment_profile_item.xml
@@ -1,39 +1,46 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:gravity="center_vertical"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:orientation="horizontal">
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+android:layout_width="match_parent"
+android:layout_height="wrap_content"
+android:minHeight="72dp">
 
     <ImageView
         android:id="@+id/personal_info_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/placeholder"/>
+        android:layout_width="@dimen/layout_list_image_size"
+        android:layout_height="@dimen/layout_list_image_size"
+        android:layout_alignParentStart="true"
+        android:layout_centerVertical="true"
+        android:layout_marginStart="@dimen/layout_list_item_margin"
+        android:layout_marginTop="@dimen/layout_list_item_margin"
+        android:contentDescription="@string/placeholder" />
 
     <LinearLayout
-        android:layout_width="wrap_content"
+        android:id="@+id/textview_container"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical">
+        android:layout_alignWithParentIfMissing="true"
+        android:layout_centerVertical="true"
+        android:layout_marginLeft="@dimen/layout_list_item_margin"
+        android:layout_marginRight="@dimen/layout_list_item_margin"
+        android:layout_toEndOf="@+id/personal_info_icon"
+        android:orientation="vertical"
+        android:paddingBottom="@dimen/layout_list_item_margin"
+        android:paddingTop="@dimen/layout_list_item_margin">
 
         <TextView
             android:id="@+id/personal_info_main_text"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textSize="@dimen/profile_text_main"
-            android:layout_marginLeft="@dimen/text_margin"
-            android:layout_marginRight="@dimen/text_margin"
-            android:textAppearance="?attr/textAppearanceListItem" />
+            android:text="@string/placeholder"
+            android:textAppearance="@style/TextAppearance.AppCompat.Body2" />
 
         <TextView
             android:id="@+id/personal_info_sub_text"
             android:layout_width="wrap_content"
-            android:textSize="@dimen/profile_sub_text"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="@dimen/text_margin"
-            android:layout_marginRight="@dimen/text_margin"
-            android:textAppearance="?attr/textAppearanceListItem" />
+            android:ellipsize="end"
+            android:maxLines="2"
+            android:text="@string/placeholder"
+            android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
     </LinearLayout>
-</LinearLayout>
+</RelativeLayout>
