+++ app/src/main/res/values-ar/strings.xml
@@ -25,4 +25,7 @@
     <string name="preference_sort_by_name_asc">حسب الاسم أسك</string>
     <string name="preference_sort_by_name_desc">حسب الاسم ديسك</string>
     <string name="preference_sort_by_frequency">حسب الشعبية</string>
+    <string name="description_frequency">تردد</string>
+    <string name="description_created_at">خلق</string>
+    <string name="description_package_name">صفقة</string>
 </resources>
